requirements.txt
GitPython==3.1.44
h5py==3.12.1
matplotlib==3.8.3
natsort==8.4.0
numpy==2.2.4
pandas==2.2.3
peft==0.9.0
regex==2023.12.25
scikit_learn==1.4.0
scipy==1.15.2
seqlbtoolkit==0.6.17
spacy==3.7.2
torch==2.4.0
transformers==4.46.0
trl==0.10.1
vllm==0.6.3.post1
wget==3.2
spacy-alignments==0.9.1
pytextspan==0.5.7
rich==13.7.1
